requirements.txt
aiosqlite==0.16.1
asyncpg==0.22.0
click==7.1.2
h11==0.12.0
iso8601==0.1.14
PyPika==0.44.1
pytz==2020.5
starlette==0.14.2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.4
